requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
torch-geometric>=2.3.0
stable-baselines3>=2.0.0
gymnasium>=0.29.0

# Web Framework & API
flask>=2.3.0
flask-cors>=4.0.0
marshmallow>=3.19.0
pydantic>=2.0.0

# Task Queue & Caching
celery>=5.3.0
redis>=4.5.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.4.0
alembic>=1.11.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0
seaborn>=0.12.0

# Model Management & Explainability
mlflow>=2.4.0
shap>=0.42.0

# Model Serving (Optional)
fastapi>=0.100.0
uvicorn>=0.22.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
pyjwt>=2.8.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
beautifulsoup4>=4.12.0
